Fix #23: Invoke SAMPLE_Function as part of Process command
The CFE framework sample/lab applications did not have any
example use-cases where a _library_ function was invoked.

For the FSW verification aspect, it is important to do
this at least in one place as it verifies that the
run-time linking is working properly, not just for
functions in the core image but also for dynamically
loaded libraries.

The unit test framework also needs to support this use
case, to demonstrate that it can link against the stub
provided by the library for the local UT.

This just simply adds a call to the "SAMPLE_Function"
provided by sample_lib to demonstrate this.