Fix #97, check event spec string
The fully-rendered strings are not entirely consistent, affected
by many external factors outside the control of the test cases.

There was even a warning in the comment against doing this, but
it was included nonetheless as an example of how it can be done.
This now keeps the original example but keeps it in the comment only,
and converts the actual test to follow the recommended practice of
only testing the spec/format string, not the fully-rendered string.

This should be much more stable going forward, and should work with
both CCSDS v1 and v2 configs.